abstract common functions/components
