+++ v2/features/farms/FarmsTableRowHeader.tsx
@@ -1,7 +1,6 @@
 import { FC, HTMLAttributes } from "react";
 import Image from "next/image";
 import { ChevronDownIcon } from "@heroicons/react/solid";
-import { JarDefinition } from "picklefinance-core/lib/model/PickleModelJson";
 import { useTranslation } from "next-i18next";
 
 import { classNames, formatDollars } from "v2/utils";
@@ -11,6 +10,7 @@ import { CoreSelectors, JarWithData } from "v2/store/core";
 import FarmComponentsIcons from "./FarmComponentsIcons";
 import { Network } from "../connection/networks";
 import FarmAPY from "./FarmAPY";
+import ChainProtocol from "v2/components/ChainProtocol";
 
 const RowCell: FC<HTMLAttributes<HTMLElement>> = ({ children, className }) => (
   <td
@@ -23,45 +23,12 @@ const RowCell: FC<HTMLAttributes<HTMLElement>> = ({ children, className }) => (
   </td>
 );
 
-const chainProtocol = (jar: JarDefinition, networks: Network[] | undefined): JSX.Element => {
-  return (
-    <div>
-      <p className="font-title font-medium text-base leading-5 group-hover:text-primary-light transition duration-300 ease-in-out">
-        {jar.depositToken.name}
-      </p>
-      <div className="flex mt-1">
-        <div className="w-4 h-4 mr-1">
-          <Image
-            src={formatImagePath(jar.chain, networks)}
-            className="rounded-full"
-            width={20}
-            height={20}
-            layout="responsive"
-            alt={jar.chain}
-            title={jar.chain}
-          />
-        </div>
-        <p className="italic font-normal text-xs text-foreground-alt-200">{jar.protocol}</p>
-      </div>
-    </div>
-  );
-};
-
 interface Props {
   simple?: boolean;
   open: boolean;
   jar: JarWithData;
   userDillRatio: number;
 }
-
-const formatImagePath = (chain: string, networks: Network[] | undefined): string => {
-  const thisNetwork = networks?.find((network) => network.name === chain);
-  if (thisNetwork) {
-    return `/networks/${thisNetwork.name}.png`;
-  } else {
-    return "/pickle.png";
-  }
-};
 
 const FarmsTableRowHeader: FC<Props> = ({ jar, simple, open, userDillRatio }) => {
   const { t } = useTranslation("common");
@@ -79,7 +46,7 @@ const FarmsTableRowHeader: FC<Props> = ({ jar, simple, open, userDillRatio }) =>
     <>
       <RowCell className={classNames(!open && "rounded-bl-xl", "rounded-tl-xl flex items-center")}>
         <FarmComponentsIcons jar={jar} />
-        {chainProtocol(jar, networks)}
+        <ChainProtocol asset={jar} networks={networks} />
       </RowCell>
       <RowCell>
         <p className="font-title font-medium text-base leading-5">
